+++ src/fi/testbed2/DownloadParameter.java
@@ -1,7 +1,0 @@
-package fi.testbed2;
-
-import android.content.Context;
-
-public class DownloadParameter {
-	public Context context;
-}