+++ src/fi/testbed2/task/ParseAndInitTask.java
@@ -1,0 +1,88 @@
+package fi.testbed2.task;
+
+import android.app.Activity;
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.graphics.Bitmap;
+import android.graphics.drawable.BitmapDrawable;
+import android.preference.PreferenceManager;
+import fi.testbed2.*;
+import fi.testbed2.app.MainApplication;
+import fi.testbed2.data.MapImage;
+import fi.testbed2.data.ParsedHTML;
+import fi.testbed2.data.TestbedMapImage;
+import fi.testbed2.exception.DownloadTaskException;
+import fi.testbed2.result.ParseAndInitTaskResult;
+import fi.testbed2.result.TaskResultType;
+import fi.testbed2.util.HTMLUtil;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Task which parses the testbed HTML page and initializes the animation view
+ * with the latest map image.
+ */
+public class ParseAndInitTask extends AbstractTask<ParseAndInitTaskResult> {
+
+    private String url;
+
+	public ParseAndInitTask(final Context context, final Activity activity) {
+        super(activity);
+        initURL(context);
+	}
+
+    private void initURL(final Context context) {
+        // construct URL where to download content
+        SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
+        String mapType = sharedPreferences.getString("PREF_MAP_TYPE", "radar");
+        String mapTimeStep = sharedPreferences.getString("PREF_MAP_TIME_STEP", "5");
+        String mapNumberOfImages = sharedPreferences.getString("PREF_MAP_NUMBER_OF_IMAGES", "10");
+        url = activity.getString(R.string.testbed_base_url, mapType, mapTimeStep, mapNumberOfImages);
+    }
+
+    @Override
+    protected void saveResultToApplication(ParseAndInitTaskResult result) {
+        List<MapImage> latestMapImage = new ArrayList<MapImage>();
+        latestMapImage.add(result.getLatestImage());
+        MainApplication.setMapImageList(latestMapImage);
+        MainApplication.setParsedHTML(result.getParsedHTML());
+    }
+
+    @Override
+	protected ParseAndInitTaskResult doInBackground(Void... params) {
+
+        try {
+            ParseAndInitTaskResult result = new ParseAndInitTaskResult(TaskResultType.OK, "OK");
+
+            publishProgress(new DownloadTaskProgress(0, 0, true, activity.getString(R.string.progress_parsing)));
+            ParsedHTML parsed = HTMLUtil.parseHTML(url);
+            result.setParsedHTML(parsed);
+
+            publishProgress(new DownloadTaskProgress(50, 0, false, activity.getString(R.string.progress_downloading)));
+            result.setLatestImage(processLatestImage(parsed));
+
+            publishProgress(new DownloadTaskProgress(100, 0, false, activity.getString(R.string.progress_done)));
+            return result;
+        } catch(DownloadTaskException e) {
+            return new ParseAndInitTaskResult(TaskResultType.ERROR, e.getMessage());
+        }
+
+    }
+
+    private MapImage processLatestImage(ParsedHTML parsed) throws DownloadTaskException {
+
+        TestbedMapImage image = parsed.getLatestTestbedImage();
+        Bitmap bitmap = image.getDownloadedBitmapImage();
+        String localTimestamp = image.getLocalTimestamp();
+        MapImage mapImage = new MapImage(localTimestamp, new BitmapDrawable(bitmap));
+        return mapImage;
+
+    }
+
+    @Override
+    protected void onTaskEnd() {
+        activity.finish();
+    }
+
+}