+++ src/fi/testbed2/view/AnimationView.java
@@ -1,4 +1,4 @@
-package fi.testbed2;
+package fi.testbed2.view;
 
 import java.util.List;
 
@@ -12,6 +12,8 @@ import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
 import android.widget.TextView;
+import fi.testbed2.app.MainApplication;
+import fi.testbed2.data.MapImage;
 
 public class AnimationView extends View {
 
@@ -103,10 +105,10 @@ public class AnimationView extends View {
 		super.onDraw(canvas);
 		
 		String text = String.format("%1$2d/%2$2d @ ", currentFrame + 1 , frames + 1) + mapImageList.get(currentFrame).timestamp;
-		
+
 		timestampView.setText(text);
 		timestampView.invalidate();
-		
+
 		BitmapDrawable frame = mapImageList.get(currentFrame).bitmapDrawable;
 		frame.setBounds(bounds);
 		frame.draw(canvas);
@@ -184,13 +186,17 @@ public class AnimationView extends View {
 		return true;
 	}
 
+    public void refresh(Context context) {
+        this.init(context);
+    }
+
 	private void init(Context context) {
         
 		// get default frame delay
 		SharedPreferences sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context);
 		frameDelay = Integer.parseInt(sharedPreferences.getString("PREF_ANIM_FRAME_DELAY", "1000"));
 		
-        mapImageList = MyApplication.getMapImageList();
+        mapImageList = MainApplication.getMapImageList();
         
     	// Assume all frames have same dimensions
     	frameWidth = mapImageList.get(0).bitmapDrawable.getMinimumWidth();